pyproject.toml
[project]
name = "audire-flow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.3.3",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "python-multipart>=0.0.9",
    "python-docx>=1.1.0",
    "mammoth>=1.8.0",
    "pymupdf>=1.23.0",
]
